+++ app/Http/Controllers/PermissionController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StorePermissionRequest;
 use App\Http\Requests\UpdatePermissionRequest;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Spatie\Permission\Models\Permission;
 
 class PermissionController extends Controller
@@ -14,6 +15,8 @@ class PermissionController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('permission-index'), 403);
+
         return view('permissions.index')->with([
             'permissions' => Permission::all(),
         ]);
@@ -24,6 +27,8 @@ class PermissionController extends Controller
      */
     public function create()
     {
+        abort_if(Gate::denies('permission-create'), 403);
+
         return view('permissions.create');
     }
 
@@ -42,6 +47,8 @@ class PermissionController extends Controller
      */
     public function show(Permission $permission)
     {
+        abort_if(Gate::denies('permission-show'), 403);
+
         return view('permissions.show')->with([
             'permission' => $permission,
         ]);
@@ -52,6 +59,8 @@ class PermissionController extends Controller
      */
     public function edit(Permission $permission)
     {
+        abort_if(Gate::denies('permission-edit'), 403);
+
         return view('permissions.edit')->with([
             'permission' => $permission,
         ]);
@@ -72,6 +81,8 @@ class PermissionController extends Controller
      */
     public function destroy(Permission $permission)
     {
+        abort_if(Gate::denies('permission-destroy'), 403);
+
         $permission->delete();
 
         return redirect()->route('permissions.index')->with('success', 'Permiso eliminado correctamente');
